+++ CHANGES.md
@@ -15,10 +15,14 @@ All releases are available on [Anaconda.org](https://anaconda.org/conda-forge/ge
 
 ## Unreleased
 
+- {gh}`1159` Add Einsatzgemeinschaft (SGB XII); remove Einstandsgemeinschaft.
+  ({ghuser}`MImmesberger`)
 - {gh}`1154` Fix Freibetrag for pension income from all pillars for Grundsicherung im
   Alter. ({ghuser}`MImmesberger`)
 - {gh}`1155` Fix Vermögensfreibetrag for Grundsicherung im Alter.
   ({ghuser}`MImmesberger`)
+- {gh}`1160` Make `tests-with-cov` use loop vectorization to enable coverage reporting.
+  ({ghuser}`hmgaudecker`)
 
 ## v1.2 — 2026-03-19